method update

+++ src/main/java/com/github/MrPatronO/carworkshop/repositories/CarRepository.java
@@ -4,9 +4,8 @@ import com.github.MrPatronO.carworkshop.models.Car;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
 
 @Repository
-public interface CarRepository extends JpaRepository<Car, Integer> {
+public interface CarRepository extends JpaRepository<Car, Long> {
 
 }
